refactor: consolidate date utils
