+++ lib/dateUtils.ts
@@ -1,13 +1,16 @@
+const AMERICA_BOGOTA = "America/Bogota";
 const longDateFormatter = new Intl.DateTimeFormat("es-CO", {
   day: "numeric",
   month: "long",
   weekday: "long",
   year: "numeric",
+  timeZone: AMERICA_BOGOTA,
 });
 const shortDateFormatter = new Intl.DateTimeFormat("es-CO", {
   day: "numeric",
   month: "numeric",
   year: "numeric",
+  timeZone: AMERICA_BOGOTA,
 });
 
 export const getLocalLongDateString = longDateFormatter.format;
@@ -24,3 +27,12 @@ export function getWeekdayName(date: Date = new Date()): string {
 export function isValidDateString(date: any): date is string {
   return typeof date === "string" && !Number.isNaN(new Date(date));
 }
+
+export const dateParts = (date: Date) => {
+  const parts = shortDateFormatter.formatToParts(date);
+  return Object.fromEntries(
+    parts
+      .filter(({ type }) => ["year", "month", "day"].includes(type))
+      .map(({ type, value }) => [type, Number(value)])
+  ) as Record<"year" | "month" | "day", number>;
+};
